[BEAM-6627] Moved ByteMonitor, TimeMonitor and ConsoleResultPublisher to testutils
