+++ sdks/java/io/file-based-io-tests/build.gradle
@@ -30,7 +30,6 @@ dependencies {
   shadowTest project(path: ":beam-sdks-java-io-xml", configuration: "shadowTest")
   shadowTest project(path: ":beam-sdks-java-io-parquet", configuration: "shadowTest")
   shadowTest project(path: ":beam-sdks-java-test-utils", configuration: "shadowTest")
-  shadowTest project(path: ":beam-sdks-java-load-tests", configuration: "shadowTest")
   shadowTest library.java.guava
   shadowTest library.java.junit
   shadowTest library.java.hamcrest_core
